# jobnow-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JobNow API",
        "description": "API documentation for JobNow - Moroccan Job Portal Platform. This API provides endpoints for job seekers and companies to manage job offers, applications, and user profiles.",
        "contact": {
            "name": "JobNow Support",
            "email": "support@jobnow.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.jobnow.ma",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/search/autocomplete/jobs": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Autocomplete job titles",
                "description": "Returns job title suggestions based on search query",
                "operationId": "6ace43840de49ca421ef412011dac823",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query (minimum 2 characters)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autocomplete suggestions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "suggestions": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/search/autocomplete/locations": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Autocomplete locations",
                "description": "Returns city suggestions based on search query",
                "operationId": "b068ba9bb42c11b8da495c7ff6fb3e59",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query (minimum 2 characters)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location suggestions"
                    }
                }
            }
        },
        "/api/search/autocomplete/companies": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Autocomplete company names",
                "description": "Returns company name suggestions based on search query",
                "operationId": "ec7941b024c83f24213ac36945c2110f",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query (minimum 2 characters)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company suggestions"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}